requirements.txt
flask
python-vlc
opencv-python
pillow
screeninfo
qrcode[pil]
psutil
GPUtil
wmi
pywin32